Conflicts:
	ffn_bot/fanfiction_parser.py

+++ ffn_bot/reddit_bot.py
@@ -6,6 +6,7 @@ import logging
 import praw
 
 from ffn_bot import fanfiction_parser
+from ffn_bot import reddit_markdown
 from ffn_bot import ao3
 from ffn_bot import ffa
 from ffn_bot import aff
@@ -42,10 +43,13 @@ SITES = [
 # ffnbot!submissionlink      Direct-Links just for the submission-url
 CONTEXT_MARKER_REGEX = re.compile(r"ffnbot!([^ ]+)")
 
-FOOTER = "\n\nSupporting fanfiction.net (*linkffn*), AO3 (buggy) (*linkao3*), HPFanficArchive (*linkffa*), and FictionPress (*linkfp*)." + \
+FOOTER = "\n\nSupporting fanfiction.net (*linkffn*), AO3 (buggy) (*linkao3*), HPFanficArchive (*linkffa*), FictionPress (*linkfp*), AdultFanFiction (linkaff) (story ID only)" + \
     "\n\nRead usage tips and tricks  [**here**](https://github.com/tusing/reddit-ffn-bot/blob/master/README.md).\n\n" + \
     "^(**New Feature:** Parse multiple fics in a single call with;semicolons;like;this!)\n\n" + \
-    "^^**Update** ^^**7/7/2015:** ^^More ^^formatting ^^bugs ^^fixed.\n\nffnbot!ignore"
+    "^(**New Feature:** Type 'ffnbot!directlinks' in any comment to have the bot **automatically parse fanfiction links** and make a reply, without even calling the bot! Added AdultFanFiction support!)" + \
+    "\n\n^^**Update** ^^**7/11/2015:** ^^More ^^formatting ^^bugs ^^fixed. ^^Feature ^^added!\n\n^^^^^^^^^^^^^^^^^ffnbot!ignore"
+
+DRY_RUN = False
 
 
 def get_regexps():
@@ -63,15 +67,17 @@ def get_sites():
 def run_forever():
     sys.exit(_run_forever())
 
+
 def _run_forever():
     """Run-Forever"""
     while True:
         try:
             main()
+        # Exit on sys.exit and keyboard interrupts.
+        except KeyboardInterrupt:
+            raise
         except SystemExit as e:
             return e.code
-        except KeyboardInterrupt:
-            return 0
         except:
             logging.error("MAIN: AN EXCEPTION HAS OCCURED!")
             bot_tools.print_exception()
@@ -79,6 +85,7 @@ def _run_forever():
 
 
 def main():
+    global DRY_RUN
     """Basic main function."""
     # moved call for agruments to avoid double calling
     bot_parameters = get_bot_parameters()
@@ -86,6 +93,7 @@ def main():
     load_checked_comments()
     load_subreddits(bot_parameters)
 
+    DRY_RUN = bool(bot_parameters["dry"])
     while True:
         for SUBREDDIT in SUBREDDIT_LIST:
             parse_submissions(r.get_subreddit(SUBREDDIT))
@@ -96,24 +104,39 @@ def get_bot_parameters():
     """Parse the command-line arguments."""
     # initialize parser and add options for username and password
     parser = argparse.ArgumentParser()
-    parser.add_argument('-u', '--user', help='define Reddit login username')
-    parser.add_argument('-p', '--password', help='define Reddit login password')
     parser.add_argument(
-        '-s', '--subreddits', help='define target subreddits; seperate with commas')
-
-    # Can also add possibility with -s option to aquire comma separated list of subreddits
-    # then do: subs = args.subreddit.split(',')
-    # and return this list, then append/extend to Subreddit_list or default_subs
-    # and possibl transform to set to avoid duplicates.
-
-    # Alternatively we can just use docopt. (See http://docopt.org/)
+        '-u', '--user',
+        help='define Reddit login username'
+    )
+    parser.add_argument(
+        '-p', '--password',
+        help='define Reddit login password'
+    )
 
     parser.add_argument(
-        '-d', '--default', action='store_true', help='add default subreddits, can be in addition to -s')
+        '-s', '--subreddits',
+        help='define target subreddits; seperate with commas'
+    )
+
+    parser.add_argument(
+        '-d', '--default',
+        action='store_true',
+        help='add default subreddits, can be in addition to -s'
+    )
+
+    parser.add_argument(
+        '-l', '--dry',
+        action='store_true',
+        help="do not send comments."
+    )
 
     args = parser.parse_args()
 
-    return {'user': args.user, 'password': args.password, 'user_subreddits': args.subreddits, 'default': args.default}
+    return {
+        'user': args.user, 'password': args.password,
+        'user_subreddits': args.subreddits, 'default': args.default,
+        'dry': args.dry
+    }
 
 
 def login_to_reddit(bot_parameters):
@@ -147,8 +170,14 @@ def load_subreddits(bot_parameters):
 
 def check_comment(id):
     """Marks a comment as checked."""
-    global CHECKED_COMMENTS
+    global CHECKED_COMMENTS, DRY_RUN
     CHECKED_COMMENTS.add(str(id))
+
+    # This is a dry run, do not pretend we
+    # store anything in any way.
+    if DRY_RUN:
+        return
+
     with open('CHECKED_COMMENTS.txt', 'w') as file:
         for id in CHECKED_COMMENTS:
             file.write(str(id) + '\n')
@@ -158,8 +187,12 @@ def load_checked_comments():
     """Loads all comments that have been checked."""
     global CHECKED_COMMENTS
     logging.info('Loading CHECKED_COMMENTS...')
-    with open('CHECKED_COMMENTS.txt', 'r') as file:
-        CHECKED_COMMENTS = {str(line.rstrip('\n')) for line in file}
+    try:
+        with open('CHECKED_COMMENTS.txt', 'r') as file:
+            CHECKED_COMMENTS = {str(line.rstrip('\n')) for line in file}
+    except IOError:
+        bot_tools.print_exception()
+        CHECKED_COMMENTS = set()
     print('Loaded CHECKED_COMMENTS.')
     logging.info(CHECKED_COMMENTS)
 
@@ -200,11 +233,12 @@ def parse_submission_text(submission):
         additions
     )
 
+
 def parse_submissions(SUBREDDIT):
     """Parses all user-submissions."""
     print("==================================================")
     print("Parsing submissions on SUBREDDIT", SUBREDDIT)
-    for submission in SUBREDDIT.get_hot(limit=25):
+    for submission in SUBREDDIT.get_hot(limit=50):
         # Also parse the submission text.
         if not is_submission_checked(submission):
             parse_submission_text(submission)
@@ -221,7 +255,7 @@ def parse_submissions(SUBREDDIT):
                 print("Parsing comment ", comment.id, ' in submission ', submission.id)
                 try:
                     make_reply(comment.body, comment.id, comment.id, comment.reply)
-                except:
+                except Exception:
                     logging.error("\n\nPARSING COMMENT: Error has occured!")
                     bot_tools.print_exception()
                     check_comment(comment.id)
@@ -241,7 +275,11 @@ def make_reply(body, cid, id, reply_func, markers=None, additions=()):
         print('Outgoing reply to ' + id + ':\n' + reply + FOOTER)
         print('--------------------------------------------------')
         print(Style.RESET_ALL)
-        reply_func(reply + FOOTER)
+
+        # Do not send the comment.
+        if not DRY_RUN:
+            reply_func(reply + FOOTER)
+
         bot_tools.pause(1, 20)
         print('Continuing to parse submissions...')
     else:
